Rename "scan_json" to "scan".
