+++ scan_json/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod action;
 pub mod matcher;
-pub mod scan_json;
+pub mod scan;
 
 pub use action::{BoxedAction, BoxedEndAction, BoxedMatcher, StreamOp, Trigger};
 pub use matcher::{Matcher, Name, ParentAndName};
-pub use scan_json::{scan_json, ContextFrame};
+pub use scan::{scan, ContextFrame};
